[Refactor] 뷰컨트롤러 분리
